requirements.txt
anyio==3.6.2
bcrypt==4.0.1
cffi==1.15.1
click==8.1.3
colorama==0.4.6
configparser==5.3.0
cryptography==39.0.1
ecdsa==0.18.0
fastapi==0.91.0
greenlet==2.0.2
h11==0.14.0
httptools==0.5.0
idna==3.4
jose==1.0.0
mysql-connector-python==8.0.32
passlib==1.7.4
pathlib==1.0.1
protobuf==3.20.3
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.4
PyJWT==2.6.0
PyMySQL==1.0.2
python-decouple==3.7
python-dotenv==0.21.1
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.3
starlette==0.24.0
typing_extensions==4.4.0
unicorn==2.0.1.post1
uvicorn==0.20.0
watchfiles==0.18.1
websockets==10.4
